Fix clang warnings.
* src/base/ftoutln.c (FT_Outline_EmboldenXY): Fix type of
`orientation'.

* src/gxvalid/gxvcommn.c (gx_lookup_value_read): Fix signature.

* src/pcf/pcfread.c (pcf_get_encodings): Fix type of some variables.
Add cast.

* src/type1/t1load.c (parse_weight_vector): Fix cast.

+++ src/type1/t1load.c
@@ -1043,7 +1043,7 @@
       blend->default_weight_vector[n] =
       blend->weight_vector[n]         = T1_ToFixed( parser, 0 );
 
-      FT_TRACE4(( " %f", (float)blend->weight_vector[n] / 65536 ));
+      FT_TRACE4(( " %f", (double)blend->weight_vector[n] / 65536 ));
     }
 
     FT_TRACE4(( "]\n" ));
